icap-v7-master/postprocess/requirements.txt
asgiref==3.11.0
Django==4.0.2
sqlparse==0.5.5
typing_extensions==4.15.0
psycopg2-binary==2.9.10
groq==0.33.0
python-dotenv==1.1.1
requests==2.32.5
gunicorn>=20.1.0
whitenoise==6.5.0
pika==1.3.2
redis==4.5.1
django-import-export==3.3.7
